fall back to loading raw ipynb
If jupytext conversion fails (commonly due to git putting conflict
markers in the JSON), fall back to editing the raw ipynb so that we can
fix it up.

Without this change, vim will present you with an empty buffer in that
situation. If you wanted to resolve the conflict markers, you need to
toggle g:jupytext_enable and reload the file first.

Raw ipynb seems to get loaded with an extra blank line at the top, not
sure why that happens.